move all api requests into one utils file, remove bloated content, link all imports
